Logging refactor
Take the logging stuff and put it in it's own module
